Adapt snapshot tests for Python 3.7
Since Python 3.8 read-only properties are shown separately (see
https://bugs.python.org/issue36401).  As Python 3.7 is the lowest 3.x version
that will be supported in the future, dedicated snapshot files are introduced
only for this particular version.  The files without a version-specific prefix
should always correspond to the latest supported Python version, which is 3.8
at this point.

Note that the snapshot tests as implemented right now will simply fail when run
with Python 2.  As support for running genpybind using Python 2 will be dropped
soon due to Python 2's end of life status, this is not considered a defect.